requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
build==1.4.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.5
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.6
cssselect==1.3.0
dataclasses-json==0.6.7
datasets==4.5.0
dill==0.4.0
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.1
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
greenlet==3.0.3
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==0.3.28
langchain-chroma==0.2.6
langchain-community==0.3.31
langchain-core==0.3.83
langchain-google-genai==2.1.12
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
langsmith==0.4.37
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.0.2
oauthlib==3.3.1
onnxruntime==1.19.2
openai==2.30.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyarrow==21.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
PyPika==0.51.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.27.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.22.2
tomli==2.4.1
torch==2.8.0
tqdm==4.67.3
transformers==4.57.6
typer==0.23.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.39.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wikipedia==1.4.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0